cpu/esp32: disable multiheap and SPI RAM
Due instabilities in memory management,  multiheap handling is disabled with in current implementation. newlibc memory management is used instead. Since SPI RAM access requires mutliheap heandling capabilities, module esp_spi_ram is disabled for the moment. If it is tried to be used by the application, an error message is thrown during make.