cpu/atmega328p: Changed PCINT configuration
Pins PD2 and PD3 can be used as both external interrupt and as pin changed
interrupt. (PD2: INT0/PCINT18, PD3: INT1/PCINT19.) This commit removes support
for the pin changed interrupt PCINT18 and PCINT19 on the ATmega328P MCU in
order to reflect that GPIO PD2 and PD3 cannot be set up as port changed
interrupt, as RIOT will always use the external interrupt of the pin changed
interrupt if both are available.