tests/pkg_tlsf_malloc: Test TSLF is correctly initialized.
This test is currently failing because of #4490, #5796 and #12021.

When using TLSF as the system allocator it should be initialized

- Automatically, as that is what the user expects.
- Early in the boot process, since the C library mallocs internal buffers.

Failing to do so will lead to a crash as the issues and this test shows.

The test is blacklisted and will be whitelisted in the next commit with
the fix.