at86rf2xx: add phy states
This change defines PHY states for the radio. Although the internal
states differ between Basic Mode and Extended Mode operation, they have
an equivalent in PHY states (e.g RX_AACK_ON is equivalent to RX_ON).
With this it will be easier to implement Basic Mode support for this
radio