at86rf2xx: use static buffer protection
Due to the design of netdev, it's more convenient to use static buffer
protection instead of dynamic. As soon as the frame buffer is accessed,
the SFD detection is disabled until the whole recv procedure finished
(dropping or fetching from FB). This way it's not necessary to switch to
PLL_ON mode on reception.