at86rf2xx: reimplmenet state machine
The current driver SPI polls the states of the transceiver. This doesn't
prevent race conditions to happen, since these states are very volatile.
With this change the PHY states of the transceiver are represented with
the `dev->state` variable, which syncs to the device via interrupts.
Now it's possible to safely check if the device is busy without chances
of race conditions.