sys/shell: Make echoing and prompt runtime-configurable.
This adds a persistent state to the shell that stores the echo configuration
and the prompt character. Both can be changed by the user with special built
in commmands

- @eon: turn echo on.
- @eon: turn echo off.
- @ps x: make x the prompt character. "@ps" (without any char) disables
  the prompt.

This is implemented in a way that does not consume any additional RAM- it
only adds some bytes to the code. A couple of bytes are stolen from the
user supplied buffer to store this new configurations.

The default state us given by the same compile time defines as before so
this change is backwards compatible.