sys/shell: refactor tokenizer code.
The tokenizer (the code that breaks up the line given to the shell into
strings to create argv) was quite a messy piece of code. This commit
refactors it into a more traditional state-machine based parser.

This fixes the issues with quote handling exposed by the recently
introduced test.

Not only is the code clearer now, it is also smaller:

In the SAMR-21, before:

```
   text	   data	    bss	    dec	    hex	filename
  10292	    140	   2612	  13044	   32f4	tests/shell/bin/samr21-xpro/tests_shell.elf
```

After:

```
   text    data     bss     dec     hex filename
  10268	    140	   2612	  13020	   32dc	tests/shell/bin/samr21-xpro/tests_shell.elf
```

Further commits can bring this number down even further, but for now the
focus in on clarity.