pyterm: use clean output by default
Do not prepend the output with the timestamp and the loglevel (always
INFO) by default.
This gives a different behavior with the other RIOT_TERMINAL. Also, CI
uses `picocom` by default so does not have timestamp which can make a
test fail locally when it works in CI. Or work in native and fail on a
board as native does not have this decoration.

Also it will help with test when you need to get the output of the node
and not from your terminal program.

The previous behavior can be achieved by using the environment variable

    PYTERMFLAGS='--format "%(asctime)s - %(levelname)s # %(message)s"

So for example

    PYTERMFLAGS='-f "%(asctime)s - %(levelname)s # %(message)s"' \
        RIOT_TERMINAL=pyterm BOARD=samr21-xpro make -C tests/shell/ flash term