makefiles/utils/strings: optimize `rest` function.
The rest function gets the "tail" of a list and is often used to define
recursive list processing functions.

It uses make's wordlist function which requires an upper bound, which was
being computed using the "words" function. Replacing this by a very high
number (2**31 - 1) makes it much faster and will never be an issue (the
build machine will probably run out of memory before).

Note that GNU Make does not like values larger than 2**31.