examples/default: PoC user/password login for for the terminal
This is a very rough proof of concept showing how a simmple user/password
prompt can be used at the serial terminal to protect the shell.

To be effective, this requires a shell that can be exited, in order to be
able to log off.

The login prompt has a built in delay between attempts (1 second at lest, 7
seconds each three failed attempts).

This won't work quite right in native because the serial/pty handling is
different there (we are not turning off the OS' line bufering and that gets
in the way, as well as ctrl-d)

I get the best experience using miniterm.py:

```
miniterm.py --eol LF /dev/ttyACM0 115200
```