pyterm: Add a --device-timeout option
When pyterm is used on boards that use CDC-ACM stdio (ie. provide their
own USB device), that device is not available immediately after
flashing, and a `make flash term` is likely to fail.

This adds a --device-timeout option to pyterm that will repeat pyterm's
device opening step until the timeout is exceeded (in which case it will
fail as it had before), or the device can be opened.

Boards that use such a configuration will want to set
PYTERMFLAGS += --device-timeout=5 or another suitable value depending on
the board's expected time to go from bootloader to operational.