boards/common/arduino-atmega: remove special case for jiminy-mega256rfr2
The board is deprecated, no need to ignore this board anymore here.

Having the `arduino` and `periph_pwm` features is now required for all
boards using 'arduino-atmega' (as it was except for that board).

If this should change in the future, it should be defined either in each
arduino board, in another board common, or per CPU_MODEL.