drivers/mrf24j40: don't loop in mrf24j40_reset_state_machine()
When hooking up the mrf24j40 to a bluepill board, the driver would
always get stuck on init.

Debugging revealed that it would get stuck in the mrf24j40_reset_state_machine()
function because it expects the RFSTATE to have a special value after reset.

However, the data sheet does not mention this in section 3.1 Reset.
Waiting 192µs should be enough - the value of the RFSTATE is not specified.

The Linux driver also does not wait for the RFSTATE register.

And even without the loop, the driver is functioning fine.