squash! dist/pythonlibs/testrunner: make reset after term optional
For some board `make reset` is only possible if a serial connection
is not already open or its execution might disrupt it. This
casuses some tests to fail since before running a test the board
is reset. The reset is not performed is TESTRUNNER_RESET_SYNC env
variable is set.