boards: add mcb2388
The MCB2388 Evaluation Board is a development board for the LPC2388 MCU.

LPC2388 and LPC2387 are very simmilar, the only difference is added USB Host
support.
Since this is not used yet, I chose to just pretend it's an lpc2387.

So far, only the two UARTs, LEDs and LCD display are configured.