periph/gpio: add gpio_init_low() / gpio_init_high() helper functions
Provide a way to init a GPIO as output with a set level.
Some platforms can to this on init already and will always just set
the GPIO to LOW which can be undesireable.

For all other platforms, provide a fall-back implementation.

The functions are also convenient when de-initializing a previously
otherwise configured Pin, e.g. when disconnecting the power on an external
sensor through a transistor and wanting to avoid current leaking through a
high idle UART TX line.