boards/atmega328p: Fixed xtimer frequency
The ATmega328p on a breadboard can have various frequencies depending on
the fuse settings and whether and which crystal is connected. Thus, finding a
fixed xtimer frequency that can be configured for every CPU core clock with the
available prescalers is not possible. Therefore, the fixed frequency is replaced
by a frequency depending on the CPU clock speed. For CPU clock frequency of
more than 4 MHz a prescaler of 64 will be chosen (same as on the Arduino Uno),
for 4 MHz and less a prescaler of 8 will be chosen.