at: Wait 20ms before a new command operation
According to the ublox cellular modem AT reference manual,
a delay of -at least- 20ms is required before starting a new
command operation so that the modem has time to send the stored
URC messages to the host. It is also important to note that
this delay is inserted just BEFORE draining the AT device so as
to receive and process possible URCs. Another point to keep in
mind is that this delay is put in functions which constitute
a whole AT command, i.e. sending the command and waiting for the
answer. The remaning functions like at_send_bytes or at_recv_bytes
do not have this delay. The users, who will use those functions in
their own implementation, are advised to respect this timing.