Revert "drivers/hd44780: add short delay after each character"
This reverts commit cf01c743a89e17d78037fcc77f32b8daa859845e.

Adding an unexplained delay seemed wrong in the first place, but it fixed
the display on the MCB2388.

Turns out the display was erroneously operating in 8-bit mode due to the
uninitialized flag register.
Why the delay helped here I don't know.

But with #12634 fixing this, this hack is not needed anymore.