semtech-loramac: adapt to sx127x watchdog code
The Semtech package has some hooks for TX and RX timeout (in case of
lost interrupt or if something goes wrong with the hardware). We enable
the sx127x watchdog here to keep the same behavior.
Note that although the package sets RX and TX timeout to dynamic values,
a LoRaWAN packet will neved exceed a time on air of 3 seconds (default
of SX127x watchdog timers). So, the timeout value of the stack is not
used and we just rely on the driver settings.