gnrc_lwmac: don't use constant as argument for RTT callback
The RTT callback for a super-frame cycle uses the `arg` pointer to set
the message value that then is handed to the lwmach thread.
However, in both instances the timer is scheduled the constant
`GNRC_LWMAC_EVENT_RTT_WAKEUP_PENDING` is provided.
This means the argument is not really necessary.

Same as e1f81f0 for gomac.

fixes #12869