tests: provide test for `vfs`+`fmt` corner-case
On `native` when using `vfs` the `fmt` print functions do not work.
This is because the `fmt` module uses the `write()` system call which
is bend by the `native_vfs` module to use `vfs_write()`. However,
`native` does not use a `stdio` module to print. Instead, it just
writes to the hosts standard I/O directly. As such, STDIN, STDOUT, and
STDERR are never initialized for `vfs` so `vfs` does not recognize
`STDIN_FILENO`, `STDOUT_FILENO` and `STDERR_FILENO` as valid file
descriptors.

This test case showcases this bug.