tinydtls: sock_dtls: fix deadlock in sock_dtls_recv()
In the current implementation of `sock_dtls_recv()` a timeout message
event is armed to fire after `timeout` us. If `sock_udp_recv()` takes
a long time to return with a valid message (res > 0), then the timeout
message might fire **before** the timer is removed with
`xtimer_remove()`. This leads to a message in the `mbox` that is never
removed. After a few iterations, this results in a full `mbox` and
hence a deadlock of `sock_dtls_recv()` (particularly: in the `_read()`
event of tinydtls).