boards/mega-xplained: fix ADC line defintions
For the ATmega platform the ADC lines are not defined according to the GPIOs. Instead, they are defined from 0 to ADC_NUMOF-1, which in turn represents the bit in the corresponding ADC registers.  The mapping from the line number to the corresponding ADC register/bit combination is done implicitly by the periph/adc implementation.