Added SPI bus support for the lpsxxx driver.
This commit adds SPI support to the lpsxxx driver.
The driver now supports I2C as well as SPI, but the unique thing is
that both SPI and I2C can be supported within a single build.

The periph/comms_bus.h and periph/comms_bus.c implementation supplies
the glue for pivoting between I2C and SPI busses, during
run-time. These files are meant to be generic and free to be used by
other peripherals that support both I2C and SPI.