murdock: don't run static tests
static tests are already executed by Travis.
Failing Travis will already block the merge, so we don't have to run
those tests in Murdock again.

This has two benefits:

 - slightly decreased CI build times
 - Murdock failures will only be caused by build errors.

The last point is my main motivation for this patch.
It often happens that Murdock failures are dismissed during review if
a PR is not squashed yet.
Even if there are still build failures, they are not visible at a glance.

Travis will already fail when the PR is not yet squashed, so Murdock doesn't
have to check this again and can now only warn about serious issues.