boards/same54-xpro: don't source peripheral clocks from main clock
Use the dedicated 48 MHz clock as a source for the peripheral clocks.
This was already done for I2C to allow it to work despite the 120 MHz
main clock.

Not running the peripherals off the main clock will allow for dynamic
re-clocking of the main clock in the future, without affecting the
operation of the peripherals.

It also gives more flexibility to the main clock selection in general.