tests/periph_rtc: test sub-second precision
Adapted from @benemorius PR #10763.
Add mircosecond precision test to RTC.
Adapt automated python test to match test API changes.
Assert max of 100 ms  error.

This is needed to test that an RTC alarm fires at the desired time and not with an arbitrary offset up to +/-1 second.
It also tests that rtc_set_time() behaves equivalently correctly, which is what I needed to test when I wrote this.