sys/auto_init: add gnrc_netif_manual_init
Using this module allows to delay the initialisation of network interfaces.

When the gnrc_netif_manual_init module is used, network interfaces will only
be set up when the application manually calls auto_init_gnrc_netif().

This is useful to conserve power, e.g. the application would only power up the
network interfaces when it needs them.

As an example, at86rf215 consumes 3.6 mA when left unconfigured and 60 mA when
ready to receive on both interfaces.
When in deep sleep, the device consumes 1nA.

A battery powered sensor might use hibernate mode for maximum power saving.
That means a reset is triggered on wake-up.
The sensor might wake up every 5 minutes to take a measurement and store it
in some none-volatile memory or backup RAM, but only send out the data once
every hour.

So we don't want to automatically start the radio interfaces each time we wake
up, only when the application decides to send data.

Before going to sleep, the application will put the network interfaces into deep
sleep by calling

    gnrc_netif_t* netif = NULL;
    netopt_state_t state = NETOPT_STATE_SLEEP;
    while ((netif = gnrc_netif_iter(netif))) {
        gnrc_netapi_set(netif->pid, NETOPT_STATE, 0, &state, sizeof(netopt_state_t));
    }

As long as init is not touching the radio chip, it will remain in deep sleep
until the application calls auto_init_gnrc_netif() again.