boards/nucleo-l152re: fix uart1 pinout config
According to the pinout the rx and tx pin need to be the other way around.
Though the swapped tx and rx pins allow the uart to work in loopback, it is still not correct.
The PA10 functions as an RX pin even though PA10 states TX pin in the periph_conf.
With this fixed the pullup for the rx is configured correctly and noise doesn't trigger the line.

This change sets the .tx_pin from PA10 to PA9 and the .rx_pin from PA9 to PA10.