drivers/at86rf215: adapt default config to same54-xpro EXT1
Previously the default config was for connecting the REB215-XPRO
module to EXT3 on same54-xpro.

Since now more SPI buses are configured on same54-xpro and the default
config also serves as a compile test for boards with only one SPI bus,
connect the module to the (new) first SPI bus by default.