boards/atmega1284p: Default programmer, fix doc
The AVR Dragon, previously the cheapest debugger/programmer for AVR, is out of
production. The default debugger for AVR devices has been changed to the Atmel
ICE, the now cheapest debugger option. The commit updates the documentation
accordingly.

Additionally the default programmer is changed to the Atmel ICE for consistency.
The doc is updated accordingly.